Update patch set 1
Patch Set 1:

The change seems fine, but I'm curious about which devices exhibit a problem with the current arrangement and are fixed by this change. 

I believe both the sound playing and the stopping of recording are asynchronous to an extent, so rearranging may not do much. But if you have evidence of this fixing things on some devices, I'm happy to take the fix.

Patch-set: 1