Update patch set 1
Patch Set 1:

They are async, and we would not write in blood that the order after change is guaranteed, however we have solved the problem in one of our not released products this way.

Apart from this, I think to do it in this order is more logic, first stop recording then play sound.

Patch-set: 1