Update patch set 1
Patch Set 1:

James, please reread the commit description. This does not change the code one single bit about what code actually is used, the exact same codepaths are still used.

It is only a first step to deobfuscate the current code, and the effect is that you don't need to define -D__arm__ when compiling it on other architectures as you have to right now. Defining -D__arm__ when compiling for another architecture is a portability disaster waiting to happen (yes I ran into problems with that when testing the code on other platforms).

Patch-set: 1