Update patch set 1
Patch Set 1:

I see. But even if the OMX component doesn't allow it, this makes the code more complete, if someone for some reason were to use it via something else than the current OMX wrapper. I think this is a sensible extra feature addition, even if it is not of use via the OMX wrapper.

Patch-set: 1