Update patch set 1
Patch Set 1:

I didn't talk to PV, but I tested the encoder in a setup of mine, with non-multiple-of-16 dimensions, and it worked just fine (no valgrind warnings either). Without this patch, it also worked, but the output video contained gray borders.

This patch in itself doesn't change anything in the normal, multiple-of-16 dimensions case that is used by the OMX component.

Patch-set: 1