Update patch set 1
Patch Set 1: Do not submit

we probably do not want to change the behavior, since if we could not seek, we do not want to falsely tell the apps that the seek is completed and successful. As a result of the fasle completion notification, the app would misled and mistakenly update the playback/streaming position.

Patch-set: 1
Label: Code-Review=-2