Update patch set 1
Patch Set 1:

Ok, if you prefer the fix going into MPEG2TSWriter, I'll try to fix it there.

I can't assume there is no other codec that does this, but I have never seen any such except for this one.

The startcodes shouldn't necessarily be added by the wrapped software encoder - that's up to the API of the software encoder. The API of this one is clear - it doesn't add any startcodes at all - which is fine.

The SoftAVCEncoder wrapper however already does adds startcodes to the CODECCONFIG data (the software encoder returns the SPS and PPS as two separate buffers in two separate function calls, SoftAVCEncoder puts them into one buffer with startcodes before each of them). Therefore I find it inconsequential to add startcodes to one kind of buffer but not to all of them.

Additionally, if passing data from the encoder straight on to a decoder without writing it into a container inbetween, I'm not sure if any existing H264 decoders (either HW or the google SW ones) can handle the data that this encoder outputs, without adding the startcodes. I can build an example test case of this to show what I mean.

Patch-set: 1