Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

There is no need to modify SampleTable::findSyncSampleNear
with this patch it no longer finds the closest sample correctly.

Here is an example from a sample table that contains the following samples:

Index   Time
[...]
1252    1252000
1253    1253000
1254    1254000
[...]

Without this patch a request for the closest index for time 1252620 returns index 1253 (1253000)

With the patch a request for the closest index with time 1253670 returns index 1253 (1253000). 
It should return index 1254 since 1254000 is closer to 1253670

There is a problem with SampleTable::findSyncSampleNear. For kFlagClosest
it does not return the closes sample, but rather what you would expect kFlagAfter to return.

The binary search in findSampleAtTime and findSyncSampleNear

while (left < right) {
    [...]
}

returns the *highest* (rightmost) of the two closest values if no value matches. This is taken into account in findSampleAtTime 
where the second sample to evaluate is chosen as mSampleTimeEntries[closestIndex - 1]

In findSyncSampleNear the second sample is chosen as mSyncSamples[left + 1] (note +1 vs -1)

The reason that this works in ICS is that left is adjusted directly after the search:

if (left > 0) {
    --left;
}

( rows 638 to 640 in https://android.googlesource.com/platform/frameworks/base/+/refs/heads/ics-mr1-release/media/libstagefright/SampleTable.cpp )

This patch works by modifying the search in findSyncSampleNear to return the lowest, instead of the highest, value as expected in the rest of the code.
But since the change doesn't need to be applied to findSampleAtTime i think its better to adapt findSyncSampleNear to work with the current search.

I suggest to either reintroduce the "--left" in findSyncSampleNear or to adjust the second sample to "mSyncSamples[left - 1]" instead of mSyncSamples[left + 1] , not to modify "while (left < right)"

Patch-set: 1