Create change
Patch-set: 1
Change-id: I7d1b94cffbfe7c38ca094834dedbc92a58855e20
Subject: Bug fix for the MediaPlayer::prepare() api. For an MP3 http source when a check for ID3 tags is executed for the purpose 	of checking for gapless playback info, 	if ID3v2 tags aren't present, then a check is done for ID3v1 tags. 	This results in a read command that asks the cache to jump to the 	end of the file, and subsequently make an extra http call to request 	those bytes. For a streamed source, this causes the file to not be 	downloaded automatically when MediaPlayer::prepare() is called, and 	causes stuttering and extra buffering time to be needed when start() is 	finally called. The solution is to ignore the ID3v1 tags as the gapless info would never exist 	there, and only check for ID3v2 tags.
Branch: refs/heads/master
Commit: 10eeb17f60fea49a41bd8a2b5c4397737ae6f42c