Update patch set 3
Patch Set 3:

The problem is that even if the sources do not emit stale data after seek, the notified position is taken from already decoded data being rendered.

The seektime somehow needs to be propagated to the renderer. In the RTSP case the first access unit timestamped after the seek will not be decoded and queued to the renderer until there is enough buffered data on both channels, which might take a while.

Maybe it would be possible to just stop position notifications from the renderer while seeking, but that would still be a very similar patch to this one.

Patch-set: 3
Reviewer: Zoran Jovanovic <1010077@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0