Update patch set 1
Patch Set 1: Code-Review-2

Sorry, but this is not the right way to fix this.
We can't change the fundamental allocation behavior of get_memory - many HALs depend on having multiple memory buffers allocated at once from this callback, and this change would make it only have one buffer available at a time, silently freeing up earlier buffers which are likely still in use.

However, the problem is real; the emulator camera HAL does not release recording frames properly in its releaseRecordingFrame method.  

The right fix is to track the IMemrory buffers created in CallbackNotifier::onNextFrameAvailable in a queue, so that their data pointers can be matched up to the pointers returned to CallbackNotifier::releaseRecordingFrame, and then freed.

Patch-set: 1
Label: Code-Review=-2