Update patch set 4
Patch Set 4:

Unrolling is a little restrictive at this time.

Filter sizes are

160*2*8
160*3*8
160*4*8
147*2*8
147*3*8
147*4*8

Since we do 8 VMAC (vmlal) at a time

this becomes

gcd(160*[2,3,4] and 147*[2,3,4]) VMACS = 1

there is no common factor among these six rates for a common unroll rate.

that being said, it's still doable, but would complicate the code without reducing the register usage in this particular area.

One way to do this is to do factors of 8 followed by factors of 4, then by 2 and then by 1.  The other way is to have multiple versions of the inner loop with different unroll rates.  Both of these I think should be next steps after register optimization, as this inner loop code has been shown to be quite efficient based on clock metrics.  Also, if we decide to add more input sample rate support, it would further complicate this analysis.

Patch-set: 4