Update patch set 1
Patch Set 1:

The problem with https://android-review.googlesource.com/#/c/49660/ is that it assumes STREAMINFO is 042 bytes long. I have seen files that violate this by adding padding at the end of the STREAMINFO block and plays fine in other players, so I believe we don't need to be so strict about it.

Patch-set: 1
Reviewer: Pawit Pornkitprasan <1011499@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0