Update patch set 1
Patch Set 1: Code-Review-1

Thanks for spotting. I wrote my solution for JB and there that problem does not exist. But you are right, there is a corner case when you play a non variant playlist with query string. Not very comon, but your solution works in that case too so lets go for it.

Patch-set: 1