Update patch set 2
Patch Set 2:

I wouldn't say it should. This is basically the same thing as if you've loaded http://server/foo/bar.html?a=b&c=d in a browser, and you're following an <a href="baz.html"> link within that page - the query string is unique to the original page, and if one is expected in the next one it will be part of the link (or in this case, in the playlist itself).

Patch-set: 2
Label: Verified=0