Update patch set 1
Patch Set 1: Code-Review+1

Honestly, I'm not sure how we normally handle this. In general, we use the same compiled code for userdebug or user builds, so an #if-style compile time statement won't work; it would need to be some kind of runtime check.

I *thought* verbose logging was stripped out on release builds, so we could change these to VERBOSE logging and rely on the system to do the filtering.

Sorry I can't be of more help. I agree with the spirit of this change.

Patch-set: 1
Reviewer: Nick Kralevich <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1