Update patch set 1
Patch Set 1:

Fair enough.  When I started this patch, I thought it was maybe just inconsistencies between toolchains and that including the template param was more correct.  I've looked over the latest C++ standard, and it looks like this is completely valid for the last couple years.  I also checked clang for reference and it compiles things like this fine too.

Patch-set: 1