Update patch set 1
Patch Set 1:

Sure - completely disabling the possibility for resampling in the fast path also makes sense. Now it's so long since I dug into this and debugged it so I don't remember off-hand at which level it needs to be disabled, but the fast path flag needs to be cleared earlier in the stack (OpenSL ES, AudioTrack, AudioFlinger) somewhere. Do you have a patch ready somewhere that I could test that would remove the fast path flag earlier in this case?

Patch-set: 1