Update patch set 1
Patch Set 1:

So just to clarify my previous comment - making AudioFlinger not support resampling at all in the fast path won't solve the full problem, because that's actually what is already happening. OpenSL ES doesn't know whether resampling is supported in the fast path (or doesn't know what the HW sample rate is? don't remember the detail right now), but since everything else matches, it sets the fast path flag. The AudioTrack class does the same, it knows the HW sample rate but doesn't know whether AudioFlinger supports resampling in the fast path or not, so it keeps the fast track path set and doesn't allocate any bigger buffers since the resampling won't be done at that level. Finally within AudioFlinger it realizes that fast path was requested but the sample rate doesn't match and resampling in the fast path isn't enabled - and thus disables the fast path and does resampling at this level, but with too small buffers.

So this is what my patch fixes, given that all the earlier layer passed the fast path flag through to this level so that AudioFlinger has to try to do resampling, make sure the buffers are big enough.

If we're going to disable this case altogether, the earlier layers (probably AudioTrack) needs to verify the sample rate and clear the fast path flag before invoking AudioFlinger, since the case when AudioFlinger rejects the fast path flag that was requested isn't handled too well. (If you close and reopen the same track it works better, because then AudioTrack remembers that fast path was requested but failed and thus tries without fast path the next time - and then everything works well.)

Patch-set: 1