Update patch set 2
Patch Set 2:

A test web page that showcases what this fixes is http://albin.abo.fi/~mstorsjo/hls-redirect.html - try loading this in e.g. Safari, and it will play properly, but it fails to play in Chrome on current Android devices. With this patch it should play back correctly.

Patch-set: 2