Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

I'm OK with the change, but I don't understand the motivation for it ... the commit message says what was done, but not why it was done.  What was wrong with using size_t as the size of the parameter block?

Patch-set: 1
Reviewer: Glenn Kasten <1004395@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1