Update patch set 1
Patch Set 1:

Sorry for not being more detailed. 

We were facing memory corruption on 64-bit system as a 32-data was being accessed as size_t*, with size_t being 64-bit on LP64.

From what we could see, the parameters would fit in a 32-bit value even on a 64-bit machine. In addition, the amount of changes required is less when size_t in certain functions were changed back to uint32_t. Hence we changed the size_t to int32_t.

Patch-set: 1