Update patch set 1
Patch Set 1:

Get parameter set of functions had formal parameter type as size_t*. The actual parameter passed was uint32_t*. The functions then stored the parameter at the address passed.
 
This worked fine on a 32-bit machine as size_t and uint32_t are of same size. But on LP64, we end up with memory corruption. 

Hope the problem is clear now.

Patch-set: 1