Update patch set 1
Patch Set 1:

I still don't understand completely, but what I know so far is that libstagefright fails to build in 32-bit due to various problems in it's dependencies, and that then causes a lot of other things to fail to build.   This CL helps to isolate what doesn't build, so that we can still build as much as possible in 32/64.  The current blacklist in build/core is much too broad.

Patch-set: 1