Update patch set 1
Patch Set 1:

The issue really is that frameworks/base (libandroid_runtime.so) refuses to compile because of its dependency on libmedia (which in turn depends on libstagefright). I was hoping we could have some record of why these modules don't compile, though it looks like you might have covered that in your spreadsheets.

Patch-set: 1