Update patch set 2
Patch Set 2: -Code-Review

goog/master has some of these changed downstream to suffix of __unused (the solution that is to be used for 'C' files, the standard is to comment out for 'C++' files). I will have to split this into two CLs, one for the ones that pass cleanly, and another for the one for the set that fail downstream.

Patch-set: 2