Update patch set 3
Patch Set 3:

The 'clean' cherry pick, roughly 40% of the changes had to be tossed as they resulted in regressions and patch stutter (BIG MESS, because of downstream work).

The last 3 files resulted in conflicts requiring manual intervention or regression in code.

We want this synchronization in all the trees as it has caused some pain for us, I expect that our OEM customers will also experience the pain. I chose to take a bit of an engineering burden to get as much of this patch upstream in AOSP as possible.

*one* file in this set would cause some pain for some other 64-bit fixes I have blocked until this is upstream. I had discussed this with Colin and we felt that I had the choice of cherry picking the one file, or taking an engineering look at cherry picking this whole set. I chose the later as I felt it would make a few engineer's lives easier in the long run.

Patch-set: 3