Update patch set 1
Patch Set 1:

This change makes sure that MediaCodec does not crash, if trying to play back an interlaced H264 sample with direct rendering into a Surface. Decoding in that setup fails, and this failure leaves the ACodec/MediaCodec object pair in an inconsistent state.


The qualcomm decoder doesn't like the NativeWindow that is set up:
E/OMX-VDEC-1080P( 2539): Insufficient sized buffer given for playback, expected 3145728, got 3137536
E/OMXNodeInstance( 2539): OMX_UseBuffer failed with error -2147479547 (0x80001005)
E/ACodec  ( 9261): registering GraphicBuffer 0 with OMX IL component failed: -2147483648

This error is reported and handled as intended, but the error handling codepath in ACodec for errors in port reconfiguration shuts down the codec (see the comment about "This is technically not correct, but appears to be the only way to free the component instance." in ACodec.cpp), leaving the ACodec/MediaCodec object pair in an inconsistent state. When trying to stop/release the MediaCodec object, this then triggers asserts about unexpected parameters and/or inconsistent state.

Patch-set: 1
Reviewer: Martin Storsjö <1002068@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0