Update patch set 1
Patch Set 1:

I tested this patch this evening and it certainly seems to fix the crash originally reported here:  https://android-review.googlesource.com/#/c/96172/.

Just one small comment.  I notice that there is one variable in the code that is defined as unsigned long and not ULong and hence not affected by this change.

SoftMPEGEncoder.cpp:724 defines

unsigned long modTimeMs = 0;

A pointer to this variable is then passed to the 4th argument of PVEncodeVideoFrame which expects a ULong*, i.e., a uint32_t*, on line 727.  This doesn't seem to cause any problems, but it might be better to update the type of this variable.

Patch-set: 1