Update patch set 1
Patch Set 1:

Could you change only the forward declarations to match existing class-or-struct in the definitions? That should be a simpler change to verify.
I understand that we can also change any definition to struct or class and add private or public before the first member, but I am assuming some reasons when those definitions were written as a class or struct.
Thanks.

Patch-set: 1
Label: Code-Review=0