Update patch set 1
Patch Set 1: Code-Review-1

This is one of those issues that we are considering, but for the moment, MediaMuxer does not accept codec-specific data in buffers (we modeled this after MediaExtractor). This must be passed in when adding the track. Note: there is an outstanding issue where the MPEG4-video csd is not passed into MediaFormat, so one cannot create an mp4 file with MPEG-4 video content.

Patch-set: 1
Reviewer: Lajos Molnar <1021605@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1