Update patch set 1
Patch Set 1:

(2 comments)

> Haojun, thanks for the submission.  Is there a test case that shows
 > this problem?  Please note that MediaPlayer::decode is calling
 > MediaPlayerService::decode through Binder.

Andy, thank you for your reply. No, I don't have a test case, but I do have a use case. In our company's source code, there is a use case where MediaPlayer::decode is calling MediaPlayerService::decode directly, not through a binder. (They are in the same process, for the reason of performance, somebody in our company decided to do SoundPool effect in the media server, but I'm not sure, really.)

But a binder should just be an implementation detail, right? The point of RPC is to make it like a local procedure call, now when we call it locally, we found we are closing the fd twice...

Sorry for my ill English, please feel free to contact me. Also see my inline comments.

Patch-set: 1