Update patch set 1
Patch Set 1:

Note that these are already part of the public API in the current NDK releases. Changing to an actual symbol (like for NdkMediaFormat) won't work for the current Android 5.0 out there, so that would only be for the next platform release (and would add antother ABI difference to keep track of, for users).

Patch-set: 1