Update patch set 1
Patch Set 1:

Not much of a reason from the AOSP perspective, but I occasionally build this code standalone outside of AOSP as well, and it did build fine with e.g. pre-2013 MSVC versions, while I now noticed that it no longer built. I know that supporting such setups isn't a priority for AOSP, but in this case, keeping the code C89-compatible doesn't really cost much effort, at least not fixing it occasionally if it is broken.

Patch-set: 1