Update patch set 1
Patch Set 1: Code-Review-1

Re-configuration of an existing AudioTrack is a "convenience" feature that could be implemented in a layer above AudioTrack.  We are trying to move in the direction of having AudioTrack as minimal as feasible, and adding new conveniences at a higher level.

Patch-set: 1
Reviewer: Glenn Kasten <1004395@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1