Update patch set 1
Patch Set 1:

If we cannot determine from the playlist that we are switching to an audio only stream (see M3UParser::getTypeURI()), the streamMask will still contain video as well. This means that when going FROM audio only TO audio/video we would use the last video segment as start time, which would be wrong, since it will be at least one (probably several) segment(s) behind audio.

Patch-set: 1