Update patch set 1
Patch Set 1:

Woops my wording in the message is weird. When libvpx is updated, the build will fail *unless* this is merged first. The current code supports both versions of the define, so there is no reason to wait and merge this at the same time as the update.

Patch-set: 1