Update patch set 3
Patch Set 3:

> Uploaded patch set 3.

I have added some more fixes. Swapped port reconfig and reinit so that, port reconfig is not called in the case of error. Codec context also needs to be set to null at the beginning and when a  init fails.
I have also removed the clip you uploaded.(Google drive is what is typically used).

Patch-set: 3
Label: Verified=0