Update patch set 2
Patch Set 2:

Why is AUDIO_RESAMPLER_UP_RATIO_MAX so large anyway? The lowest sampling rate you might encounter in real life is on the order of a couple of kiloHertz. Are we ever going to upsample that to over a hundred megaHertz?

Patch-set: 2
Reviewer: Marco Nelissen <1001525@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0