Update patch set 1
Patch Set 1: Code-Review-2

Unfortunately, BLOB and YUV_420_888 are not the only software-readable formats that the camera supports, and this change would prevent all other formats from working.

The software encoder should be setting its desired color format as the default consumer format, instead.  The camera service will honor the default, as long as it's not RGBA.

Patch-set: 1
Label: Code-Review=-2