Update patch set 2
Patch Set 2:

I fully expect to be regressed a few times before I have a clever thought about what the nature of the regressions are and how to avoid them.  I think for the most part, the surfaces I touch are where you talk to other processes, and those are small (for now).

As our infra gets better, we'll be able to get mail about regressions as we write tests around the camera stack.

Patch-set: 2