Update patch set 1
Patch Set 1:

> Can you explain why it's not sufficient to give an extra boost to the current locale?

This change is using system locale in AutoDetect class and that's why we can trust system locale more in CharacterEncodingDetector.
So just boosting confidence in CharacterEncodingDetector is not sufficient.

But as you said, our code is using old code. It is, maybe, bad...

As a background, we got requirement from several countries to show ID3 tag info correctly. So we implemented them in old platform. Now you have re-designed, but it's not sufficient for us. Mainly Thai language support is necessary in addition to your new design.

Is it possible to add Thai language support in your CharacterEncodingDetector?

Patch-set: 1