Update patch set 1
Patch Set 1: Code-Review-1

We found out the change were not required to fix the actual stream end issue. This change was pushed assuming we need to extend the offload related check for other compressed format as well. But as eric pointed out this is not actually needed.

Patch-set: 1
Label: Code-Review=-1