Update patch set 1
Patch Set 1:

I do believe it's a good idea to take the device locale into account somehow. Maybe like I suggested earlier to give a boost to a matching encoding that ICU detected (e.g. if the device locale is Japanese, and ICU detects Shift-JIS as a possible encoding, then boost the confidence for that)
If that doesn't work for some reason, e.g. because ICU doesn't actually think the correct encoding could be a match, then we'll need to take a  look at what's going on under the hood.

Do you have any samples of files that are misdetected?

Patch-set: 1