Update patch set 1
Patch Set 1:

Thanks! I will open an internal bug to track this. Then we can verify the fix.

You said "the bug is only reproduced on tablet". Does that mean you tried on other device, eg. Nexus Player, and did not see the bug. Or does it mean that you did not try other devices.

Patch-set: 1