Update patch set 1
Patch Set 1:

data_offset is an actual offset in the file. The only way I can see for it to end up negative is if you had a file that was 16 exabytes in size and had an atom that extended past the 64-bit limit. I'll add a check to make it explicit though.

Patch-set: 1
Label: Verified=0