Update patch set 1
Patch Set 1:

I don't see that kind of warning yet, although it is implementable.
To turn on that kind of warning and make it an error would be equivalent to adding "explicit" keyword to all such constructors.

I think having "explicit" keyword "optional" is good.
It is easy to add "explicit" to every constructor, but when necessary we can leave some constructor non-explicit.

Patch-set: 1