Update patch set 1
Patch Set 1:

Thank you for the stack trace.

This is still very odd to me, unfortunately.  BasicClient::disconnect is being called because the vector of clients to evict created on line 1528 is being destroyed as out of scope on line 1563, which is expected.

However, line 1555 already calls BasicClient::disconnect on all members of the vector, and disconnect has a guard in it to prevent attempting to disconnect twice:

CameraService.cpp, line 1925:

   void CameraService::BasicClient::disconnect() {
      if (mDisconnected) {
          return;
      }
      mDisconnected = true;
      ..

which should prevent the call to removeByClient.  Do you have modifications to BasicClient::disconnect in your source tree, compared to aosp-marshmallow?

Patch-set: 1