Update patch set 1
Patch Set 1:

Chien-Yu recalled that we'd seen a similar stack trace back in January, which was fixed by the following commit:
https://android-review.googlesource.com/#/c/194920/2

In short, in certain circumstances, the first disconnect() call in evictClientIdByRemote does not succeed, leading to a deadlock opportunity on the vector destruct.  The above CL ensures disconnect cannot fail.

Please cherry-pick that fix into your tree, and please let me know if it does not resolve the issue.

Patch-set: 1