Update patch set 3
Patch Set 3:

Thanks, Glenn.
This is multithread problem that can't be produce easily.
But the problem is very certain. We can make sure it just only review the source code.

This bug occured on very early version, and be fixed on android L with patch:
https://android.googlesource.com/platform/frameworks/av/+/63f6ffb996a7ffc9

But, It was broken because of https://android.googlesource.com/platform/frameworks/av/+/9b030df%5E!/

Patch-set: 3