Update patch set 3
Patch Set 3:

It is hard to produce this bug. We tried but are not able to do.
It is wild pointer problem. You should let it malloc other memory at the time between Visualize was delete and mCapturerThread exited.

But it is surely a problem that you can make sure from code.

Patch-set: 3