Update patch set 2
Patch Set 2:

The native MTP server thread crashes in getObject() when attempting to send a file in "USB for charging" state (before selecting MTP from the notification) because there are no storages.  This fixes the crash.

However, I'm wondering whether the MTP server should be running at all in "USB for charging" state.  The only functional difference I can see between the server running and not is that things like mtp-detect currently work but don't if the server isn't running.  In both cases file access is not possible.  I have a patch that will disable starting the MTP server, and I'll upload it shortly in case it's desirable.

Patch-set: 2