Update patch set 2
Patch Set 2:

Right, so this is very likely an inlining issue of L_mult into the loop.

Relying on using *2 instead of <<=1 is a mistake, as that can also be fixed to vectorise and you'll have the same problem again. And I'm not sure how useful it is to help with the case. Inlining cases are hard. :(

I'm assuming "coder" remains intact (ie. not inlined somewhere else), one thing we could do is to get both loops active, compile and dump in a bug in LLVM. At least we'll be able to know if the code has undefined behaviour or they're just different.

I still want to catch the bug, but it would be good to get you guys going before that, since I'm predicting this will take a while...

Just a comment: L_mult says it's the same as L_shl(A*B, 1), but it repeats the same checks (and might get it wrong). Why not use L_shl inside L_mult?

So, another alternative is, instead of re-writing the loop, just change L_mult and see if it has the same problem.

Patch-set: 2