Update patch set 7
Patch Set 7:

I've fixed configure so that it has no effect when called when the config is already open -- which I believe is the result we want. It is expected that UsbDeviceManager will send intents for Connected and Configured after start has been called, so additional configures aren't necessarily an error, they can just be ignored. As for the lock method, I agree that it's simpler and works for exclusion, but for case 3 having conditions allows start to wait till configure is finished even if it is called before configure, which is possible in some situations.

Patch-set: 7