Update patch set 2
Patch Set 2:

(2 comments)

we can wait for O dessert to have this patch, we will manage it as an AOSP patch for N.
If you agree to port on master and solve the conflict, i am ok. The split you did make really sense since Audio flinger can benefic of the enum to string translation, glad to listen this.

Patch-set: 2